Made changes suggested by belbix in pair review.
